+++ app/src/main/java/com/example/gsyvideoplayer/DetailPlayer.java
@@ -198,7 +198,7 @@ public class DetailPlayer extends AppCompatActivity {
                 })
                 .setGSYVideoProgressListener(new GSYVideoProgressListener() {
                     @Override
-                    public void onProgress(int progress, int secProgress, int currentPosition, int duration) {
+                    public void onProgress(long progress, long secProgress, long currentPosition, long duration) {
                         Debuger.printfLog(" progress " + progress + " secProgress " + secProgress + " currentPosition " + currentPosition + " duration " + duration);
                     }
                 })
